requirements.txt
# Multi-Agent Log Analysis System with CrewAI + Google Gemini
# Core framework
crewai==0.11.2
pydantic==2.12.5

# Google Gemini LLM
google-generativeai==0.5.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
python-dateutil==2.8.2
langchain==0.1.20
langchain-google-genai==0.0.5
